requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
loader==2017.9.11
loguru==0.7.2
pyTelegramBotAPI==4.15.4
python-dotenv==1.0.1
requests==2.31.0
soupsieve==2.5
urllib3==2.2.0
peewee==3.17.1
pydantic~=2.6.2
python-telegram-bot-calendar==1.0.5
